Update: charactorView 추가
